requirements.txt
asgiref==3.2.10
boto3==1.14.56
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
django-storages==1.10
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
uWSGI==2.0.19.1
drf-yasg==1.17.1
django-ckeditor==5.9.0